add section sleccion in editor
