+++ src/repo/mod.rs
@@ -1,7 +1,9 @@
 mod articles;
 mod authors;
+mod feed;
 mod sections;
 
 pub use articles::ArticlesRepo;
 pub use authors::AuthorsRepo;
-pub use sections::FeedRepo;
+pub use feed::FeedRepo;
+pub use sections::SectionsRepo;
