+++ web/static/css/article.css
@@ -256,9 +256,12 @@
   top: 50%;
   left: 50%;
   transform: translate(-50%, -50%);
-  padding: 20px;
   border: 1px solid #ccc;
   box-shadow: 0 4px 8px rgba(0, 0, 0, 0.2);
+}
+
+#credentials-dialog .form {
+  padding: 20px;
 }
 
 .edit-link {
