+++ web/static/css/editor.css
@@ -1,42 +1,52 @@
 @import url("../../variables.css");
 
 .save-button {
-    background-color: var(--color-link);
-    color: var(--color-white);
-    border: none;
-    padding: 10px 20px;
-    font-size: 16px;
-    cursor: pointer;
-    border-radius: 5px;
+  background-color: var(--color-link);
+  color: var(--color-white);
+  border: none;
+  padding: 10px 20px;
+  font-size: 16px;
+  cursor: pointer;
+  border-radius: 5px;
 }
 
 .form {
-    max-width: 100%;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    gap: 15px;
-    margin-top: 20px;
+  max-width: 100%;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 15px;
+  margin-top: 20px;
 }
 
 .form * {
-    max-width: 400px;
+  max-width: 400px;
 }
 
 input {
-    margin-bottom: 10px;
-    width: 100%;
-    padding: 10px;
-    border: 1px solid var(--color-border-secondary);
-    border-radius: 5px;
-    font-size: 16px;
+  margin-bottom: 10px;
+  width: 100%;
+  padding: 10px;
+  border: 1px solid var(--color-border-secondary);
+  border-radius: 5px;
+  font-size: 16px;
 }
 
 textarea {
-    width: 100%;
-    padding: 10px;
-    border: 1px solid var(--color-border-secondary);
-    border-radius: 5px;
-    font-size: 16px;
-    resize: vertical;
+  width: 100%;
+  padding: 10px;
+  border: 1px solid var(--color-border-secondary);
+  border-radius: 5px;
+  font-size: 16px;
+  resize: vertical;
+}
+
+select {
+  width: 100%;
+  margin: 0 auto;
+  padding: 16px;
+  border: 1px solid var(--color-border-secondary);
+  background-color: var(--color-background);
+  border-radius: 5px;
+  margin-bottom: 10px;
 }
